[UI] Phase 5.b.1: Create shared modal primitives (#18751)
Establishes ui/components/shared/Modal/ as the canonical modal foundation
that the 5 follow-on migration sub-issues (#18752-#18756) will compose
against.

New primitives (each <200 lines):
- Modal.tsx — base: title, size, actions, onClose, header/footer composition
- ConfirmModal.tsx — wraps Modal: title, message, confirm/cancel
- InfoModal.tsx — wraps Modal: title, body (markdown-capable)
- FormModal.tsx — wraps Modal: react-hook-form & RJSF integration
- useModal.ts — imperative open/close hook
- index.ts — barrel re-export

The pre-existing legacy file at ui/components/shared/Modal/Modal.tsx
(8.3 KB RJSF helper) is preserved as LegacyRJSFModal.tsx for back-compat
during the migration. Same pattern for Information/InfoModal.tsx ->
Information/LegacyInfoModal.tsx.

Sistent's Modal / ModalBody / ModalFooter / ModalStyledHeader form the
underlying material primitive; the shared layer adds the consistent prop
shape, size tokens, footer composition, and useModal hook.

No existing modal consumers are migrated in this PR — that's the scope
of #18752-#18756.

Closes #18751
Refs #18661 #18654
