+++ ui/eslint.config.js
@@ -253,8 +253,8 @@ const legacyInlineStyleOffenders = [
   'components/general/CreateDesignBtn.tsx',
   'components/shared/ErrorBoundary/ErrorBoundary.tsx',
   'components/shared/Modal/ConnectionModal.tsx',
-  'components/shared/Modal/Information/InfoModal.tsx',
-  'components/shared/Modal/Modal.tsx',
+  'components/shared/Modal/Information/LegacyInfoModal.tsx',
+  'components/shared/Modal/LegacyRJSFModal.tsx',
   'components/general/TipsCarousel.tsx',
   'components/general/error-404/index.tsx',
   'components/layout/Header/Header.tsx',
